// src/models/event-item.interface.ts
export interface EventItem {
    eventname: string;
    owner: any;
    uid: string;
    // lat: number;
    // long: number;
    date: string;
    time: string;
    map: string;
    posttime: any;
    detail: string;
    category: string;
    avatar: any;
    keyid: string;
    timestamp: string;
    // eee: string;
}